# Distinct Aggregate vectorization