--- a/src/include/processor/result/pattern_creation_info_table.h
+++ b/src/include/processor/result/pattern_creation_info_table.h
@@ -25,9 +25,8 @@ public:
 
     PatternCreationInfo getPatternCreationInfo(const std::vector<common::ValueVector*>& keyVectors);
 
-    uint64_t matchFTEntries(std::span<const common::ValueVector*> flatKeyVectors,
-        std::span<const common::ValueVector*> unFlatKeyVectors, uint64_t numMayMatches,
-        uint64_t numNoMatches) override;
+    uint64_t matchFTEntries(std::span<const common::ValueVector*> keyVectors,
+        uint64_t numMayMatches, uint64_t numNoMatches) override;
 
 private:
     uint8_t* tuple;