--- a/src/planner/operator/logical_aggregate.cpp
+++ b/src/planner/operator/logical_aggregate.cpp
@@ -1,5 +1,6 @@
 #include "planner/operator/logical_aggregate.h"
 
+#include "binder/expression/aggregate_function_expression.h"
 #include "binder/expression/expression_util.h"
 #include "planner/operator/factorization/flatten_resolver.h"
 #include "planner/operator/schema.h"
@@ -31,18 +32,23 @@ void LogicalAggregate::computeFlatSchema() {
 f_group_pos_set LogicalAggregate::getGroupsPosToFlatten() {
     auto [unflatGroup, flattenedGroups] =
         FlattenAllButOne::getGroupsPosToFlatten(getAllKeys(), *children[0]->getSchema());
-    // Flatten aggregates if they are from a different group than the unflat key group
+    // Flatten distinct aggregates if they are from a different group than the unflat key group
+    // Regular aggregates can be processed when unflat, but distinct aggregates get added to their
+    // own AggregateHashTable and have the same input limitations as the aggregate groups
     if (unflatGroup != INVALID_F_GROUP_POS) {
         for (const auto& aggregate : aggregates) {
+            auto funcExpr = aggregate->constPtrCast<binder::AggregateFunctionExpression>();
             auto analyzer = GroupDependencyAnalyzer(false /* collectDependentExpr */,
                 *children[0]->getSchema());
             analyzer.visit(aggregate);
             for (const auto& group : analyzer.getRequiredFlatGroups()) {
                 flattenedGroups.insert(group);
             }
-            for (const auto& group : analyzer.getDependentGroups()) {
-                if (group != unflatGroup) {
-                    flattenedGroups.insert(group);
+            if (funcExpr->isDistinct()) {
+                for (const auto& group : analyzer.getDependentGroups()) {
+                    if (group != unflatGroup) {
+                        flattenedGroups.insert(group);
+                    }
                 }
             }
         }