--- a/src/planner/plan/append_aggregate.cpp
+++ b/src/planner/plan/append_aggregate.cpp
@@ -10,9 +10,7 @@ void Planner::appendAggregate(const expression_vector& expressionsToGroupBy,
     const expression_vector& expressionsToAggregate, LogicalPlan& plan) {
     auto aggregate = make_shared<LogicalAggregate>(expressionsToGroupBy, expressionsToAggregate,
         plan.getLastOperator());
-    appendFlattens(aggregate->getGroupsPosToFlattenForGroupBy(), plan);
-    aggregate->setChild(0, plan.getLastOperator());
-    appendFlattens(aggregate->getGroupsPosToFlattenForAggregate(), plan);
+    appendFlattens(aggregate->getGroupsPosToFlatten(), plan);
     aggregate->setChild(0, plan.getLastOperator());
     aggregate->computeFactorizedSchema();
     aggregate->setCardinality(cardinalityEstimator.estimateAggregate(*aggregate));