--- a/src/processor/operator/aggregate/hash_aggregate.cpp
+++ b/src/processor/operator/aggregate/hash_aggregate.cpp
@@ -213,6 +213,11 @@ void HashAggregateLocalState::init(HashAggregateSharedState* sharedState, Result
         auto vector = resultSet.getValueVector(pos).get();
         keyVectors.push_back(vector);
         keyDataTypes.push_back(vector->dataType.copy());
+        leadingState = vector->state.get();
+    }
+    if (leadingState == nullptr) {
+        // All vectors are flat, so any can be the leading state
+        leadingState = keyVectors.front()->state.get();
     }
     std::vector<LogicalType> payloadDataTypes;
     for (auto& pos : info.dependentKeysPos) {
@@ -228,8 +233,8 @@ void HashAggregateLocalState::init(HashAggregateSharedState* sharedState, Result
 
 uint64_t HashAggregateLocalState::append(const std::vector<AggregateInput>& aggregateInputs,
     uint64_t multiplicity) const {
-    return aggregateHashTable->append(keyVectors, dependentKeyVectors, aggregateInputs,
-        multiplicity);
+    return aggregateHashTable->append(keyVectors, dependentKeyVectors, leadingState,
+        aggregateInputs, multiplicity);
 }
 
 void HashAggregate::initLocalStateInternal(ResultSet* resultSet, ExecutionContext* context) {