… src/processor/operator/hash_join/join_hash_table.cpp
… src/processor/result/base_hash_table.cpp
… src/processor/result/pattern_creation_info_table.cpp