This involved refactoring the AggregateHashTable a bit, since the distinct tables end up having mixed flat/unflat vectors when you add the distinct key (since it always is the last column, but may be flat when there are unflat vectors, so the existing split of flat vectors first, then unflat vectors didn't really work).

The performance difference isn't huge for most queries, but it makes a much bigger difference for aggregation over complex data types where there is more overhead when calculating the hashes (the list and struct hashing functions allocate a temporary ValueVector, which is probably the main difference in the performance).

Performance (first is from Clickbench, the others were just constructed to see how similar queries perform with lists/structs as the keys):
| Query | Before | After |
| --- | --- | --- |
| `MATCH (h:hits) RETURN h.RegionID, COUNT(DISTINCT h.UserID) AS u ORDER BY u DESC LIMIT 10;` | 2.25s | 2.1s |
| `MATCH (h:hits) RETURN h.RegionID, COUNT(DISTINCT [h.UserID]) AS u ORDER BY u DESC LIMIT 10;` | 17.5s | 2.7s |
| `MATCH (h:hits) RETURN h.RegionID, COUNT(DISTINCT {"id": h.UserID}) AS u ORDER BY u DESC LIMIT 10;` | 3.7s | 2.3s |